+++ dp/_visualizer.py
@@ -431,20 +431,18 @@ class Visualizer:
                 return {"visibility": "hidden"}
             return {"visibility": "visible"}
 
-        @self.app.callback(
-            Output("test-info", "data"),
-            Input("self-test-button", "n_clicks"),
-            State("test-info", "data"),
-            State("slider", "value"),
-        )
-        def toggle_test_mode(_, info, t):
+        @self.app.callback(Output("test-info", "data"),
+                           Input("self-test-button", "n_clicks"),
+                           State("test-info", "data"), State("slider", "value"),
+                           State("test-select-checkbox", "value"))
+        def toggle_test_mode(_, info, t, selected_tests):
             """Toggles self-testing mode.
 
-            Args:
-                n_clicks (int): This callback is triggered by clicking the
-                    self-test-button component.
-                t (int): The current timestep retrieved from the slider
-                    component.
+            Populates the test queue according to what tests are selected by
+            the checkbox.
+
+            This callback is triggered by clicking the self-test-button
+            component and updates the test info.
             """
             # No tests to be performed on the last timestep.
             if t == len(values):
@@ -453,9 +451,7 @@ class Visualizer:
 
             # Turn off testing mode.
             if info["tests"]:
-                return {
-                    "tests": [],
-                }
+                return {"tests": []}
 
             # Create list of write indices for t+1.
             write_mask = t_write_matrix[t + 1]
@@ -466,40 +462,42 @@ class Visualizer:
             # arbitrarily pick the first one.
             all_reads = list(t_read_matrix[t + 1][write_mask][0])
 
-            # TODO: Populate according to radio box.
+            # TODO: Populate test_q in separate callback.
             test_q = []
 
-            # Filling out write test.
-            # The truth list is a list of indices that are written to in the
-            # next cell.
-            test_q.append({
-                "truth": all_writes,
-                "render": [],
-                "color": CellType.WRITE,
-                "expected_triggered_id": self._primary,
-                "tip": "What cells are written to in the next frame? (Click "
-                       "in any order)"
-            })
-
-            # Filling out read test.
-            test_q.append({
-                "truth": all_reads,
-                "render": [],
-                "color": CellType.READ,
-                "expected_triggered_id": self._primary,
-                "tip": "What cells are read for the next timestep? (Click "
-                       "in any order)"
-            })
-
-            # Filling out all value tests.
-            for x, y in zip(*np.nonzero(write_mask)):
+            if "What is the next cell?" in selected_tests:
+                # Write test.
                 test_q.append({
-                    "truth": [values[t + 1][x][y]],
-                    "render": [(x, y)],
+                    "truth": all_writes,
+                    "render": [],
                     "color": CellType.WRITE,
-                    "expected_triggered_id": "user-input",
-                    "tip": f"What is the value of cell ({x}, {y})?"
+                    "expected_triggered_id": self._primary,
+                    "tip":
+                        "What cells are written to in the next frame? (Click "
+                        "in any order)"
                 })
+
+            if "What are its dependencies?" in selected_tests:
+                # Read test.
+                test_q.append({
+                    "truth": all_reads,
+                    "render": [],
+                    "color": CellType.READ,
+                    "expected_triggered_id": self._primary,
+                    "tip": "What cells are read for the next timestep? (Click "
+                           "in any order)"
+                })
+
+            if "What is its value?" in selected_tests:
+                # Value tests.
+                for x, y in zip(*np.nonzero(write_mask)):
+                    test_q.append({
+                        "truth": [values[t + 1][x][y]],
+                        "render": [(x, y)],
+                        "color": CellType.WRITE,
+                        "expected_triggered_id": "user-input",
+                        "tip": f"What is the value of cell ({x}, {y})?"
+                    })
 
             return {"tests": test_q}
 
@@ -512,7 +510,7 @@ class Visualizer:
         def display_tests(info, t):
             alert = dbc.Alert(is_open=False,
                               color="danger",
-                              class_name="alert-auto position-fixed w-25")
+                              class_name="alert-auto")
             if not info["tests"]:
                 return self._show_figure_trace(main_figure, t), alert
 
@@ -534,9 +532,8 @@ class Visualizer:
             return fig.update_traces(z=z, selector=t), alert
 
         @self.app.callback(
-            Output("correct", "is_open"),
-            Output("incorrect", "is_open"),
             Output("test-info", "data", allow_duplicate=True),
+            Output("correct-alert", "children"),
             # For manually resetting clickData.
             Output(self._primary, "clickData"),
             # Trigger this callback every time "enter" is pressed.
@@ -545,13 +542,13 @@ class Visualizer:
             State("user-input", "value"),
             State("test-info", "data"),
         )
-        def validator(_, click_data, user_input, info):
-            """Tests if user input is correct."""
+        def validate(_, click_data, user_input, info):
+            """Validates the user input."""
             if not info["tests"]:
                 return dash.no_update
 
-            if ctx.triggered_id != info["tests"][0]["expected_triggered_id"]:
-                # Wrong type of input: no update.
+            test = info["tests"][0]
+            if ctx.triggered_id != test["expected_triggered_id"]:
                 return dash.no_update
 
             if ctx.triggered_id == self._primary:
@@ -564,22 +561,29 @@ class Visualizer:
                 # Enter number.
                 answer = user_input
 
-            test = info["tests"][0]
+            # The alert for correct or incorrect input.
+            correct_alert = dbc.Alert("Incorrect!",
+                                      color="danger",
+                                      is_open=True,
+                                      duration=3000,
+                                      fade=True,
+                                      class_name="alert-auto")
+
+            # If answer is correct, remove from truth and render the test
+            # values. Also updates alert.
             truths = test["truth"]
-
-            is_correct = False
-            # Check that [x, y] is a row of truths.
             if answer in truths:
-                # Remove from truth and update render the test values.
                 truths.remove(answer)
-                info["tests"][0]["render"].append(answer)
-                is_correct = True
+                test["render"].append(answer)
+                correct_alert.children = "Correct!"
+                correct_alert.color = "success"
 
-            # If all truths have been found, pop test from queue.
+            # If all truths have been found, pop from test queue.
             if not truths:
                 info["tests"].pop(0)
 
-            return is_correct, not is_correct, info, None
+            # Updates test info, the alert, and resets clickData.
+            return info, correct_alert, None
 
         @self.app.callback(
             Output(self._primary, "figure", allow_duplicate=True),
@@ -659,36 +663,13 @@ class Visualizer:
         ]
 
         alerts = [
-            dbc.Alert("Correct!",
-                      id="correct",
-                      is_open=False,
-                      color="success",
-                      duration=3000,
-                      fade=True,
-                      className="alert-auto position-fixed w-25",
-                      style={
-                          "bottom": 10,
-                          "left": 10,
-                          "z-index": 9999,
-                      }),
-            dbc.Alert("Incorrect!",
-                      id="incorrect",
-                      is_open=False,
-                      color="danger",
-                      duration=3000,
-                      fade=True,
-                      className="alert-auto position-fixed w-25",
-                      style={
-                          "bottom": 10,
-                          "left": 10,
-                          "z-index": 9999,
-                      }),
-            html.Div(id="test-instructions",
+            html.Div(id="correct-alert",
+                     className="position-fixed z-9999 w-25",
                      style={
-                         "bottom": 80,
+                         "bottom": 10,
                          "left": 10,
-                         "z-index": 9999,
                      }),
+            html.Div(id="test-instructions", className="bottom-50 z-9999 w-25"),
         ]
 
         array_annotations = dbc.Card(